NVIDIA: complete BF16 breadth and packed Math spine

+++ src/compiler/codegen/tessera_gpu_backend_NVIDIA/test/nvidia/sm120_softmax_kernel_invalid.mlir
@@ -4,11 +4,11 @@ module {
   llvm.func @bad(%x: !llvm.ptr, %o: !llvm.ptr, %rows: i64, %columns: i64)
       attributes {nvvm.kernel} {
     tile.softmax_kernel %x, %o, %rows, %columns {
-      storage = "bf16", accum = "f32", axis = -1 : i64,
+      storage = "f64", accum = "f32", axis = -1 : i64,
       exp_mode = "approx_exp2", ftz = false
     } : !llvm.ptr, !llvm.ptr, i64, i64
     llvm.return
   }
 }
 
-// CHECK: 'tile.softmax_kernel' op requires storage="f16" or storage="f32"
+// CHECK: 'tile.softmax_kernel' op requires storage="f16", storage="bf16", or storage="f32"
